backend/requirements.txt
fastapi
uvicorn
python-multipart
pdfplumber
reportlab
python-dotenv
openai
pytesseract
pdf2image
Pillow
pytesseract